attrs (including colours) now passed via rest arg
sort of breaks "symmetry" a bit I guess due to #:x, #:y, and #:win being kwargs, but I think it's better than requiring the user -- potentially an inexperienced user of curses -- to know that attributes (and one colour) are combined using bitwise-ior.
also, the way all of the attributes and colours are packed into a long int and combined using IOR is really cool, can't imagine you see stuff like that in libraries nowadays! The fact you can apply all of these to a character using foldl is also really cool.